Getting information about tariffs pages from backend
* recieving tariffs on index page, recieving prototypes on tariff page

* recieving available cars, car rent page

* map bugfix

* setting ids

* sending request with renting info

+++ backend/Carsharing/ViewModels/Admin/Car/CreateCarVM.cs
@@ -8,10 +8,10 @@ public record CreateCarVM
     public string LicensePlate { get; init; }
 
     [JsonPropertyName("latitude")]
-    public double ParkingLatitude { get; init; }
+    public decimal ParkingLatitude { get; init; }
     
     [JsonPropertyName("longitude")]
-    public double ParkingLongitude { get; init; }
+    public decimal ParkingLongitude { get; init; }
     
     [JsonPropertyName("prototype_id")]
     public int CarModelId { get; init; }
